requirements.txt
torch
torchvision
torchaudio
transformers
pandas
openpyxl
nltk
Sastrawi
pyLDAvis
google-generativeai
streamlit
bertopic
sentence-transformers